stillopen/backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
joblib
numpy
python-multipart

# Scalable Database & GIS Extensions
sqlalchemy
psycopg2-binary
geoalchemy2
python-dotenv
duckdb
pandas
pyarrow
